+++ lib/flutter_tex.dart
@@ -1,7 +1,7 @@
 library flutter_tex;
 
-export 'src/models/tex_view_child.dart';
-export 'src/models/tex_view_decoration.dart';
+export 'src/models/tex_view_container.dart';
+export 'src/models/tex_view_tex.dart';
 export 'src/utils/tex_view_rendering_engine.dart';
 export 'src/utils/tex_view_style.dart';
 export 'src/views/platform_interface.dart';